RE-1169 Make use of Docker more customisable
In most cases, it's irrelevant which base image we use for creating a
container, however an ops-fabric test requires an ubuntu:17.10
container and subsequent tests call python3.6, which is not available
in the standard ubuntu base container image we use for all tests.

This review introduces the following three new parameters:

SLAVE_CONTAINER_DOCKERFILE_REPO
SLAVE_CONTAINER_DOCKERFILE_PATH
SLAVE_CONTAINER_DOCKERFILE_BUILD_ARGS

SLAVE_CONTAINER_DOCKERFILE_REPO allows you to use the Dockerfile inside
rpc-gating (`RE`), or the project in test (`PROJECT`).

SLAVE_CONTAINER_DOCKERFILE_PATH allows you to specify the location and
name of the Dockerfile, relative to the root of rpc-gating or the clone
of the project in test.

SLAVE_CONTAINER_DOCKERFILE_BUILD_ARGS allows you to pass in custom
build args to `docker build`, and in the case of rpc-gating's
Dockerfile.standard_job you can specify which base image to use in the
build.

NOTE: We remove the systemd-detect-virt call in
rpc_jobs/unit/single_use_slave.yml since systemd does not appear to be
installed on ubuntu:17.10. Instead, we check for the existence of the
/.dockerenv file to determine if we're inside a container. This may not
be the most reliable long-term test since Docker may remove this file,
however for now it seems to work.

Issue: RE-1169