RE-1624 Add PR commenter
This commit does the following:

1. Adds a new job template that binds to pull requests and will post a
comment to the PR with workflow details.

2. Adds a commentor project for rcbops/releases (more will need to be
created for other components in a subsequent review).

3. Adds a new function common.createComponentCommenter() which is used
to create commenter projects when a new component is added to
rcbops/releases.

4. Adds a `failure-status` to the `github-pull-request` trigger in
standard_job_premerge.yml so that a failed job will have its failure
status message set to indicate how to re-run the failed job.